fix: change prefix to val__; check no atom starts with it
We now check if the prefix 'val' is in the atom name. This is because after the compilation there might be conflicts in the val predicate names.

Consider the following example: '"p a b" & val-p-a-b'

The formula is compiled as follows:

"p a b" ->  'val-p-a-b'

(:derived (val-p-a-b)        (p a b))
	      ^                  ^
	      derived-predicate  condition (ground fluent, predicate name 'p', constants 'a' and 'b')

"val-p-a-b" -> 'val-val-p-a-b'

(:derived (val-val-p-a-b)     (val-p-a-b))
	      ^                   ^
	      derived-predicate   condition (ground fluent, predicate name 'p-a-b', no constants

The prefix is changed to `val__` to make the prefix less probable to occur in an atom name.